+++ example/lib/demo.dart
@@ -1,5 +1,5 @@
-import 'package:amap_map_example/data/demos.dart';
-import 'package:amap_map_example/main.dart';
+import 'package:amap_map2_example/data/demos.dart';
+import 'package:amap_map2_example/main.dart';
 import 'package:flutter/material.dart';
 
 class DemoPage extends StatefulWidget {