+++ example/lib/pages/overlays/polyline_texture.dart
@@ -1,4 +1,4 @@
-import 'package:amap_map/amap_map.dart';
+import 'package:amap_map2/amap_map2.dart';
 import 'package:flutter/material.dart';
 
 import 'package:x_amap_base/x_amap_base.dart';