+++ ios/Classes/AMapFlutterFactory.m
@@ -1,6 +1,6 @@
 //
 //  AMapFlutterFactory.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/10/29.
 //