+++ ios/Classes/AMapViewController.h
@@ -1,6 +1,6 @@
 //
 //  AMapViewController.h
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/10/29.
 //