+++ ios/Classes/Category/FlutterMethodChannel+MethodCallDispatch.m
@@ -1,6 +1,6 @@
 //
 //  FlutterMethodChannel+MethodCallDispatch.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/16.
 //