+++ ios/Classes/Category/MAPolylineRenderer+Flutter.m
@@ -1,6 +1,6 @@
 //
 //  MAPolylineRenderer+Flutter.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/7.
 //