+++ ios/Classes/Model/AMapMarker.m
@@ -1,6 +1,6 @@
 //
 //  AMapMarker.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/3.
 //