+++ ios/Classes/Model/AMapPolygon.h
@@ -1,6 +1,6 @@
 //
 //  AMapPolygon.h
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/12.
 //