+++ ios/Classes/Model/AMapPolyline.m
@@ -1,6 +1,6 @@
 //
 //  AMapPolyline.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/6.
 //