+++ ios/Classes/OverlayController/AMapMarkerController.m
@@ -1,6 +1,6 @@
 //
 //  AMapMarkerController.m
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/3.
 //