+++ ios/Classes/OverlayController/AMapPolygonController.h
@@ -1,6 +1,6 @@
 //
 //  AMapPolygonController.h
-//  amap_map
+//  amap_map2
 //
 //  Created by lly on 2020/11/12.
 //