+++ lib/src/types/ui.dart
@@ -1,8 +1,8 @@
 import 'dart:typed_data';
 import 'dart:ui' show Color;
-import 'package:amap_map/src/compatibility/color_extensions.dart';
+import 'package:amap_map2/src/compatibility/color_extensions.dart';
 import 'package:x_amap_base/x_amap_base.dart';
-import 'package:amap_map/amap_map.dart';
+import 'package:amap_map2/amap_map2.dart';
 
 /// 地图类型
 enum MapType {